Restructure images paths
The idea is that images should be organized by the game/tool/etc it's of, not the page it's made for. Allows for using the same image with multiple pages while keeping things tidy and without needing to store duplicates (as I was doing with the Filza Favorites screenshot before).

Also added mkdocs-site-urls to allow for paths to be relative to the `docs` folder instead of the current document path if the path begins with `site:`.